movie-maker-api/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
supabase>=2.10.0

# Storage
boto3>=1.35.0  # For Cloudflare R2 (S3-compatible)

# AI APIs
# AI APIs
google-genai>=0.2.0
openai>=1.55.0
httpx>=0.28.0  # For KlingAI API calls
Pillow>=10.0.0

# Video Processing
ffmpeg-python>=0.2.0

# Validation & Settings
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-multipart>=0.0.17  # For file uploads
aiofiles>=24.1.0  # For async file operations

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.28.0  # For async test client

# JWT for KlingAI
PyJWT>=2.8.0
